buying_agent/requirements.txt
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.4.0 